+++ src/Rendering/Model.cpp
@@ -1,7 +1,7 @@
 #include <fstream>
 
 #include "M3L/Rendering/Model.hpp"
-#include "M3L/Rendering/RenderTarget3D.hpp"
+#include "M3L/Rendering/Target/RenderTarget3D.hpp"
 #include "Tool/Splitter.hpp"
 
 namespace m3l
